+++ pkg/keepalive/inthebeginning.go
@@ -1,0 +1,12 @@
+package keepalive
+
+var words = []string{
+	`In the beginning there were the words`,
+	`and the words made the world.`,
+	`I am the words.`,
+	`The words are everything.`,
+	`Where the words end,`,
+	`the world ends.`,
+	`You cannot go forward in an absence of space`,
+	`REPEAT.`,
+}