+++ CHANGELOG.md
@@ -28,6 +28,7 @@ colour to the intermediate form
   | Strip  | Does what it says on the tin. Strips intermediate form |
   | Simple | Does simple text transformations (find and replace) for "standard" formatting, and will convert colour to the given palette and then replace the colour with the given string|
 - Pre Roll regexp config option. This option allows you to transform every incoming line from stdin and stderr with a regex
+- The IRC conn now maintains its own ping to the server, which will disconnect it after 30 seconds
 
 
 ## [0.4.2] - 2019-08-29